separate dev/opt make so that they can be done individually

+++ build_llvm.sh
@@ -10,7 +10,7 @@ function fatal
 
 DIR="$( cd "$( dirname "${BASH_SOURCE[0]}" )" && pwd )"
 LLVMSRC=llvm-15.0.7.src
-RELDIR=$DIR/llvm-15-release
+OPTDIR=$DIR/llvm-15-release
 DEVDIR=$DIR/llvm-15-release+assert
 
 TARGET=$(python3 -c 'import platform; print(platform.processor())')
@@ -31,14 +31,19 @@ case $TARGET in
 esac
 echo TARGET is $TARGET
 
+
+
+
 #rm -rf cmake cmake-15.0.7.src 2> /dev/null
 #tar xf cmake-15.0.7.src.tar.xz
 #mv cmake-15.0.7.src cmake
 
 
-##########################
-# assert on
-(rm -rf $LLVMSRC && tar xf $LLVMSRC.tar.xz &&
+if [ $1 == "dev" ]; then 
+    ##########################
+    # assert on
+    echo Building DEV in $DEVDIR
+    (rm -rf $LLVMSRC && tar xf $LLVMSRC.tar.xz &&
 	cd $LLVMSRC &&
 	mkdir build && 
 	cd build &&
@@ -50,22 +55,25 @@ echo TARGET is $TARGET
 			-DLLVM_ENABLE_ASSERTIONS=On \
 			.. &&
 	(make -j8 || make -j8 || make -j4 || make) && make install)  || fatal $DEVDIR
-
-##########################
-# assert off
-(rm -rf $LLVMSRC && tar xf $LLVMSRC.tar.xz &&
+elif [ $1 == "opt" ]; then 
+    ##########################
+    # assert off
+    echo Building OPT in $OPTDIR
+    (rm -rf $LLVMSRC && tar xf $LLVMSRC.tar.xz &&
 	cd $LLVMSRC &&
 	mkdir build &&
 	cd build &&
-		CC=gcc-11 CXX=g++-11 cmake -G 'Unix Makefiles' -DCMAKE_INSTALL_PREFIX=$RELDIR \
+		CC=gcc-11 CXX=g++-11 cmake -G 'Unix Makefiles' -DCMAKE_INSTALL_PREFIX=$OPTDIR \
 			-DLLVM_INCLUDE_BENCHMARKS=OFF \
 			-DCMAKE_BUILD_TYPE=Release \
 			-DCMAKE_CXX_STANDARD=17 \
 			-DLLVM_TARGETS_TO_BUILD=$TARGET \
 			-DLLVM_ENABLE_ASSERTIONS=Off \
 			.. &&
-	(make -j8 || make -j8 || make -j4 || make) && make install) || fatal $RELDIR
-
+	(make -j8 || make -j8 || make -j4 || make) && make install) || fatal $OPTDIR
+else
+    fatal "usage: $0 {opt|dev}"
+fi
 
 echo
 echo SUCCESS
